Restructure shelley-integration-tests.hs
